storage/reflink: start(): use hardlink for snap_on_start volume
Now that the _path_clean file is never mutated - not even by resize() -
a hardlink can be used instead of a reflink to store a source volume's
_path_clean file in the consumer volume.

A reflink would take some time for large and/or fragmented volume files,
whereas a hardlink is practically free even in that case. This is also
going to allow a better approach for is_outdated().